requirements.txt
openai
langchain
langchain-community
langchain-core
langchain-chroma==0.1.2
langchain-openai
PyPDF2==3.0.1
langchainhub
qdrant-client==1.10.1
python-dotenv
langchain-qdrant
langchain-huggingface==0.0.3
fastembed
streamlit
gpt4all==2.7.0
hurry.filesize
python-dotenv
sentence-transformers
chromadb==0.5.5
fastapi
uvicorn
python-multipart
python-semantic-release
pylint
pre-commit
